FLAVIO DEUS D+ Precisa testar as classes pra ver se esta tudo nos conformes e renomear o .war para ROOT dentro do tomcat

+++ src/main/java/controller/ManterCarroController.java
@@ -112,7 +112,7 @@ public class ManterCarroController extends HttpServlet {
             request.setAttribute("estacionamentos", Estacionamento.obterEstacionamentos());
             if (!operacao.equals("Incluir")) {
                 int id = Integer.parseInt(request.getParameter("id"));
-                Carro carro = Carro.obterCarro((long) id);
+                Carro carro = Carro.obterCarro(id);
                 request.setAttribute("carro", carro);
 
             }
